requirements.txt
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
ntlm-auth==1.5.0
pycparser==2.21
requests==2.27.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
typing_extensions==4.0.1
ujson==5.1.0
urllib3==1.26.7
